Build;

+++ build/server/models/entities/access-fields.js
@@ -18,42 +18,47 @@ const users_1 = __importDefault(require("./users"));
 const accesses_1 = __importDefault(require("./accesses"));
 const helpers_1 = require("../../helpers");
 let AccessField = AccessField_1 = class AccessField {
+    static repo() {
+        if (AccessField_1.REPO === null) {
+            AccessField_1.REPO = typeorm_1.getRepository(AccessField_1);
+        }
+        return AccessField_1.REPO;
+    }
     static async create(userId, attributes) {
         const { accessId } = attributes;
         helpers_1.assert(typeof accessId === 'number', 'AccessField.create second arg should have "accessId" id property');
-        const entity = repo().create(Object.assign({}, attributes, { userId }));
-        return await repo().save(entity);
+        const entity = AccessField_1.repo().create({ ...attributes, userId });
+        return await AccessField_1.repo().save(entity);
     }
     static async find(userId, fieldId) {
-        return await repo().findOne({ where: { id: fieldId, userId } });
+        return await AccessField_1.repo().findOne({ where: { id: fieldId, userId } });
     }
     static async all(userId) {
-        return await repo().find({ userId });
+        return await AccessField_1.repo().find({ userId });
     }
     static async exists(userId, fieldId) {
-        const found = await repo().findOne({ where: { userId, id: fieldId } });
+        const found = await AccessField_1.repo().findOne({ where: { userId, id: fieldId } });
         return !!found;
     }
     static async destroy(userId, fieldId) {
-        await repo().delete({ userId, id: fieldId });
+        await AccessField_1.repo().delete({ userId, id: fieldId });
     }
     static async destroyAll(userId) {
-        await repo().delete({ userId });
+        await AccessField_1.repo().delete({ userId });
     }
     static async update(userId, fieldId, attributes) {
-        await repo().update({ userId, id: fieldId }, attributes);
+        await AccessField_1.repo().update({ userId, id: fieldId }, attributes);
         const updated = await AccessField_1.find(userId, fieldId);
         return helpers_1.unwrap(updated);
     }
 };
+AccessField.REPO = null;
 __decorate([
     typeorm_1.PrimaryGeneratedColumn(),
     __metadata("design:type", Number)
 ], AccessField.prototype, "id", void 0);
 __decorate([
-    typeorm_1.ManyToOne(
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    type => users_1.default, { cascade: true, onDelete: 'CASCADE', nullable: false }),
+    typeorm_1.ManyToOne(() => users_1.default, { cascade: true, onDelete: 'CASCADE', nullable: false }),
     typeorm_1.JoinColumn(),
     __metadata("design:type", users_1.default)
 ], AccessField.prototype, "user", void 0);
@@ -62,9 +67,11 @@ __decorate([
     __metadata("design:type", Number)
 ], AccessField.prototype, "userId", void 0);
 __decorate([
-    typeorm_1.ManyToOne(
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    type => accesses_1.default, access => access.fields, { cascade: true, onDelete: 'CASCADE', nullable: false }),
+    typeorm_1.ManyToOne(() => accesses_1.default, access => access.fields, {
+        cascade: true,
+        onDelete: 'CASCADE',
+        nullable: false,
+    }),
     typeorm_1.JoinColumn(),
     __metadata("design:type", accesses_1.default)
 ], AccessField.prototype, "access", void 0);
@@ -84,10 +91,3 @@ AccessField = AccessField_1 = __decorate([
     typeorm_1.Entity('access_fields')
 ], AccessField);
 exports.default = AccessField;
-let REPO = null;
-function repo() {
-    if (REPO === null) {
-        REPO = typeorm_1.getRepository(AccessField);
-    }
-    return REPO;
-}
